sửa db hóa đơn

+++ DAL/Model/KhuyenMai.cs
@@ -11,7 +11,7 @@ namespace DAL.Model
         public string MaKM { get; set; }
         public string TenKhuyenMai { get; set; }
       //  public DateTime NgayTao { get; set; }
-        public int ChietKhau { get; set; }
+        public double ChietKhau { get; set; }
         public DateTime NgayBatDau { get; set; }
         public DateTime NgayKetThuc { get; set; }
         public Boolean TrangThai { get; set; }
